[Refactoring] Organize by folder
- reducers, styles divided folder
