+++ src/lib/reducers/slice.test.js
@@ -27,7 +27,7 @@ import reducer, {
 const middlewares = [thunk];
 const mockStore = configureStore(middlewares);
 
-jest.mock('./services/api');
+jest.mock('../../services/api');
 
 describe('reducer', () => {
   context('when previous state is undefined', () => {
